As a safety measure, pickleSynchroniser now creates the PXPaths.PICKLES folder before synching if it does not allready exist.
Modified default value settings in pickleUpdater to lower the 
number of parameter user had to fill.

Added to date input validation to make sure the format used by 
the user is correct and that we can compare dates properly. 

Corrected bug that occasionaly occured when there was a lack 
of data in a log files. 

EX : When it contained data for 2006-07-07 from 01:30:00 to 23:00:00
     but hour 7 had no data. 

Corrected the too many files opened bug caused by having way too many
opened files with the logger system.

Also the same line does not appear numerous times in same log file as
opposed to previously comitted versions.  

Modified client validation in pickleUpdater. 
   - No longer gives exception if date given is older than last update. 
   - Now it only dismisses the clients wich were updater sooner than specified
     dates and tries to update those, if any, who needs updating.
		
Corrected loss of data that occured occasionally when data for a certain hour was spread over two files.

Removed conflicts with previous commit :-)