--------Major changes tothe stats library----------
Modified most files within the stats library as to make it 
possible for a single machine to pickle the stats for every 
other machine avaiable.  

Modified path to log files in every files
wich needs access to log files so that they now take into 
consideration that a single machine could be doing pickles 
for numerous sources.  

Modified pickle name so that they no longer include the name of the machine
where they were generated, but rather the name of the machine from wich the
log files came form. 

Modified getGraphicsForWebPages so that it uses the new 
rrd functions.  

Added generalStatsLibraryMethods file wich contains methods that are to be 
used by numerous programs in the library. 
    At this point it includes such methods :
        getPathToLogFiles
        getPathToConfigFiles
        updateConfigurationFiles
                    
Also removed these methods from files were they were used. 
Code is now leaner and easier to modify.
                                
Modified key format in PICKLED-TIMES file. Now includes name of the machine from wich
the data comes from.     

Corrected a few small errors within the statMonitor utility.

Completely modified paths to graphics as to make the cleaning of the directories easier.
Modified StatsPlotter, generateRRDGraphics, and all the web pages utilities so that they 
now use the new paths.